simple_nero/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
pydantic
google-api-python-client
google-auth
python-dotenv
pytest
pytest-asyncio
httpx
